+++ src/interfaces/ffi/tests.rs
@@ -11,6 +11,7 @@ use crate::{
         gpsw::{AbeScheme, AsBytes, Gpsw},
         policy::{attr, Policy},
     },
+    error::FormatErr,
     interfaces::{ffi::error::get_last_error, hybrid_crypto::EncryptedHeader},
 };
 
@@ -30,7 +31,9 @@ use super::hybrid_gpsw_aes::{
     h_aes_destroy_encryption_cache, h_aes_encrypt_header, h_aes_encrypt_header_using_cache,
 };
 
-unsafe fn encrypt_header(meta_data: &Metadata) -> anyhow::Result<EncryptedHeader<Aes256GcmCrypto>> {
+unsafe fn encrypt_header(
+    meta_data: &Metadata,
+) -> Result<EncryptedHeader<Aes256GcmCrypto>, FormatErr> {
     let public_key_json: Value = serde_json::from_str(include_str!(
         "../hybrid_crypto/tests/public_master_key.json"
     ))?;
@@ -107,7 +110,7 @@ struct DecryptedHeader {
 
 unsafe fn decrypt_header(
     header: &EncryptedHeader<Aes256GcmCrypto>,
-) -> anyhow::Result<DecryptedHeader> {
+) -> Result<DecryptedHeader, FormatErr> {
     let user_decryption_key_json: Value = serde_json::from_str(include_str!(
         "../hybrid_crypto/tests/fin_confidential_user_key.json"
     ))?;
@@ -166,14 +169,17 @@ unsafe fn decrypt_header(
     })
 }
 
-unsafe fn unwrap_ffi_error(val: i32) -> anyhow::Result<()> {
+unsafe fn unwrap_ffi_error(val: i32) -> Result<(), FormatErr> {
     if val != 0 {
         let mut message_bytes_key = vec![0u8; 4096];
         let message_bytes_ptr = message_bytes_key.as_mut_ptr() as *mut c_char;
         let mut message_bytes_len = message_bytes_key.len() as c_int;
         get_last_error(message_bytes_ptr, &mut message_bytes_len);
         let cstr = CStr::from_ptr(message_bytes_ptr);
-        anyhow::bail!("ERROR: {}", cstr.to_str()?);
+        return Err(FormatErr::InternalOperation(format!(
+            "ERROR: {}",
+            cstr.to_str()?
+        )));
     } else {
         Ok(())
     }
@@ -181,7 +187,7 @@ unsafe fn unwrap_ffi_error(val: i32) -> anyhow::Result<()> {
 
 unsafe fn encrypt_header_using_cache(
     meta_data: &Metadata,
-) -> anyhow::Result<EncryptedHeader<Aes256GcmCrypto>> {
+) -> Result<EncryptedHeader<Aes256GcmCrypto>, FormatErr> {
     let public_key_json: Value = serde_json::from_str(include_str!(
         "../hybrid_crypto/tests/public_master_key.json"
     ))?;
@@ -263,7 +269,7 @@ unsafe fn encrypt_header_using_cache(
 
 unsafe fn decrypt_header_using_cache(
     header: &EncryptedHeader<Aes256GcmCrypto>,
-) -> anyhow::Result<DecryptedHeader> {
+) -> Result<DecryptedHeader, FormatErr> {
     let user_decryption_key_json: Value = serde_json::from_str(include_str!(
         "../hybrid_crypto/tests/fin_confidential_user_key.json"
     ))?;
@@ -332,7 +338,7 @@ unsafe fn decrypt_header_using_cache(
 }
 
 #[test]
-fn test_ffi_hybrid_header() -> anyhow::Result<()> {
+fn test_ffi_hybrid_header() -> Result<(), FormatErr> {
     unsafe {
         let meta_data = Metadata {
             uid: vec![1, 2, 3, 4, 5, 6, 7, 8, 9],
@@ -356,7 +362,7 @@ fn test_ffi_hybrid_header() -> anyhow::Result<()> {
 }
 
 #[test]
-fn test_ffi_hybrid_header_using_cache() -> anyhow::Result<()> {
+fn test_ffi_hybrid_header_using_cache() -> Result<(), FormatErr> {
     unsafe {
         let meta_data = Metadata {
             uid: vec![1, 2, 3, 4, 5, 6, 7, 8, 9],