+++ src/interfaces/pyo3/generate_gpsw_keys.rs
@@ -6,7 +6,7 @@ use crate::{
         gpsw::{scheme::GpswMasterPrivateKey, AsBytes, Gpsw},
         Engine,
     },
-    interfaces::policy::{AccessPolicy, Attribute, Policy},
+    interfaces::policy::{AccessPolicy, Attribute, Policy, PolicyAxis},
 };
 
 #[pyfunction]
@@ -33,7 +33,7 @@ pub fn generate_user_private_key(
 ) -> PyResult<Vec<u8>> {
     let master_private_key =
         GpswMasterPrivateKey::<Bls12_381>::from_bytes(&master_private_key_bytes)?;
-    let policy = serde_json::from_slice(policy_bytes.to_vec().as_slice())
+    let policy = serde_json::from_slice(&policy_bytes)
         .map_err(|e| PyTypeError::new_err(format!("Policy deserialization failed: {e}")))?;
     let access_policy = AccessPolicy::from_boolean_expression(&access_policy_str)?;
 
@@ -47,6 +47,28 @@ pub fn generate_user_private_key(
 }
 
 #[pyfunction]
+pub fn generate_policy(
+    policy_axis_bytes: Vec<u8>,
+    max_attribute_value: usize,
+) -> PyResult<Vec<u8>> {
+    let policy_axis: Vec<PolicyAxis> = serde_json::from_slice(&policy_axis_bytes)
+        .map_err(|e| PyTypeError::new_err(format!("Policy Axis deserialization failed: {e}")))?;
+    let mut policy = Policy::new(max_attribute_value);
+    for axis in policy_axis.iter() {
+        let attrs = axis
+            .attributes
+            .iter()
+            .map(std::ops::Deref::deref)
+            .collect::<Vec<_>>();
+        policy = policy.add_axis(&axis.name, &attrs, axis.hierarchical)?;
+    }
+    let policy_bytes = serde_json::to_vec(&policy)
+        .map_err(|e| PyTypeError::new_err(format!("Error serializing policy: {e}")))?;
+
+    Ok(policy_bytes)
+}
+
+#[pyfunction]
 pub fn rotate_attributes(attributes_bytes: Vec<u8>, policy_bytes: Vec<u8>) -> PyResult<Vec<u8>> {
     let attributes: Vec<Attribute> = serde_json::from_slice(&attributes_bytes)
         .map_err(|e| PyTypeError::new_err(format!("Error deserializing attributes: {e}")))?;
@@ -56,8 +78,6 @@ pub fn rotate_attributes(attributes_bytes: Vec<u8>, policy_bytes: Vec<u8>) -> Py
     for attr in attributes.iter() {
         policy.rotate(attr)?;
     }
-    let new_policy_bytes = serde_json::to_vec(&policy)
-        .map_err(|e| PyTypeError::new_err(format!("Error serializing policy: {e}")))?;
-
-    Ok(new_policy_bytes)
+    serde_json::to_vec(&policy)
+        .map_err(|e| PyTypeError::new_err(format!("Error serializing policy: {e}")))
 }