+++ src/interfaces/pyo3/hybrid_gpsw_aes.rs
@@ -3,6 +3,8 @@
 #![allow(clippy::unused_unit)]
 // Wait for `wasm-bindgen` issue 2774: https://github.com/rustwasm/wasm-bindgen/issues/2774
 
+use std::convert::TryInto;
+
 use cosmian_crypto_base::{
     hybrid_crypto::Metadata,
     symmetric_crypto::{aes_256_gcm_pure::Aes256GcmCrypto, SymmetricCrypto},
@@ -29,7 +31,7 @@ use crate::{
 type PublicKey = <Gpsw<Bls12_381> as AbeScheme>::MasterPublicKey;
 type UserDecryptionKey = <Gpsw<Bls12_381> as AbeScheme>::UserDecryptionKey;
 
-pub const MAX_CLEAR_TEXT_SIZE: usize = 1_usize << 30;
+pub const MAX_CLEAR_TEXT_SIZE: usize = 1 << 30;
 
 /// Extract header from encrypted bytes
 #[pyfunction]
@@ -43,12 +45,8 @@ pub fn get_encrypted_header_size(encrypted_bytes: Vec<u8>) -> PyResult<u32> {
     }
 
     //
-    // Recover header from `encrypted_bytes`
-    let mut header_size_bytes = [0; 4];
-    header_size_bytes.copy_from_slice(&encrypted_bytes[0..4]);
-    let header_size = u32::from_be_bytes(header_size_bytes);
-
-    Ok(header_size)
+    // Recover header size from `encrypted_bytes`
+    Ok(u32::from_be_bytes(encrypted_bytes[..4].try_into()?))
 }
 
 #[pyfunction]
@@ -84,6 +82,38 @@ pub fn encrypt_hybrid_header(
     ))
 }
 
+fn internal_decrypt_hybrid_header(
+    user_decryption_key_bytes: &[u8],
+    encrypted_header_bytes: &[u8],
+) -> PyResult<(Vec<u8>, Vec<u8>)> {
+    //
+    // Parse user decryption key
+    let user_decryption_key = UserDecryptionKey::from_bytes(user_decryption_key_bytes)?;
+
+    //
+    // Finally decrypt symmetric key using given user decryption key
+    let cleartext_header: ClearTextHeader<Aes256GcmCrypto> =
+        core_decrypt_hybrid_header::<Gpsw<Bls12_381>, Aes256GcmCrypto>(
+            &user_decryption_key,
+            encrypted_header_bytes,
+        )
+        .map_err(|e| {
+            PyTypeError::new_err(format!(
+                "Error decrypting header:
+    {e}"
+            ))
+        })?;
+
+    let metadata = cleartext_header.meta_data.to_bytes().map_err(|e| {
+        PyTypeError::new_err(format!(
+            "Serialize metadata failed:
+    {e}"
+        ))
+    })?;
+
+    Ok((cleartext_header.symmetric_key.to_bytes(), metadata))
+}
+
 /// Decrypt with a user decryption key an encrypted header
 /// of a resource encrypted using an hybrid crypto scheme.
 #[pyfunction]
@@ -91,55 +121,17 @@ pub fn decrypt_hybrid_header(
     user_decryption_key_bytes: Vec<u8>,
     encrypted_header_bytes: Vec<u8>,
 ) -> PyResult<(Vec<u8>, Vec<u8>)> {
-    //
-    // Check `user_decryption_key_bytes` input param and store it locally
-    if user_decryption_key_bytes.is_empty() {
-        return Err(PyTypeError::new_err("User decryption key is empty"));
-    }
-
-    //
-    // Check `encrypted_bytes` input param and store it locally
-    if encrypted_header_bytes.len() < 4 {
-        return Err(PyTypeError::new_err(
-            "Size of encrypted value cannot be less than 4!",
-        ));
-    }
-
-    //
-    // Parse user decryption key
-    let user_decryption_key = UserDecryptionKey::from_bytes(&user_decryption_key_bytes)?;
-
-    //
-    // Finally decrypt symmetric key using given user decryption key
-    let cleartext_header: ClearTextHeader<Aes256GcmCrypto> =
-        core_decrypt_hybrid_header::<Gpsw<Bls12_381>, Aes256GcmCrypto>(
-            &user_decryption_key,
-            &encrypted_header_bytes,
-        )
-        .map_err(|e| PyTypeError::new_err(format!("Error decrypting header: {e}")))?;
-
-    let metadata = cleartext_header
-        .meta_data
-        .to_bytes()
-        .map_err(|e| PyTypeError::new_err(format!("Serialize metadata failed: {e}")))?;
-
-    Ok((cleartext_header.symmetric_key.to_bytes(), metadata))
+    internal_decrypt_hybrid_header(&user_decryption_key_bytes, &encrypted_header_bytes)
 }
 
 /// Symmetrically Encrypt plaintext data in a block.
 #[pyfunction]
 pub fn encrypt_hybrid_block(
     symmetric_key_bytes: Vec<u8>,
-    uid_bytes: Option<Vec<u8>>,
-    block_number: Option<usize>,
+    uid_bytes: Vec<u8>,
+    block_number: usize,
     plaintext_bytes: Vec<u8>,
 ) -> PyResult<Vec<u8>> {
-    //
-    // Check `plaintext_bytes` input param
-    if plaintext_bytes.is_empty() {
-        return Err(PyTypeError::new_err("Plaintext value is empty"));
-    }
-
     //
     // Parse symmetric key
     let symmetric_key =
@@ -147,67 +139,52 @@ pub fn encrypt_hybrid_block(
             .map_err(|e| PyTypeError::new_err(format!("Deserialize symmetric key failed: {e}")))?;
 
     //
-    // Parse other input params
-    let uid = uid_bytes.map_or_else(Vec::new, |v| v.to_vec());
-    let block_number_value = block_number.unwrap_or(0);
-
-    //
     // Encrypt block
-    let ciphertext =
-        core_encrypt_hybrid_block::<Gpsw<Bls12_381>, Aes256GcmCrypto, MAX_CLEAR_TEXT_SIZE>(
-            &symmetric_key,
-            &uid,
-            block_number_value as usize,
-            &plaintext_bytes,
-        )
-        .map_err(|e| PyTypeError::new_err(format!("Error encrypting block: {e}")))?;
-
-    Ok(ciphertext)
+    core_encrypt_hybrid_block::<Gpsw<Bls12_381>, Aes256GcmCrypto, MAX_CLEAR_TEXT_SIZE>(
+        &symmetric_key,
+        &uid_bytes,
+        block_number,
+        &plaintext_bytes,
+    )
+    .map_err(|e| PyTypeError::new_err(format!("Error encrypting block: {e}")))
 }
 
+fn internal_decrypt_hybrid_block(
+    symmetric_key_bytes: &[u8],
+    uid_bytes: &[u8],
+    block_number: usize,
+    encrypted_bytes: &[u8],
+) -> PyResult<Vec<u8>> {
+    //
+    // Parse symmetric key
+    let symmetric_key =
+        <Aes256GcmCrypto as SymmetricCrypto>::Key::try_from_bytes(symmetric_key_bytes.to_vec())
+            .map_err(|e| PyTypeError::new_err(format!("Deserialize symmetric key failed: {e}")))?;
+
+    //
+    // Decrypt block
+    core_decrypt_hybrid_block::<Gpsw<Bls12_381>, Aes256GcmCrypto, MAX_CLEAR_TEXT_SIZE>(
+        &symmetric_key,
+        uid_bytes,
+        block_number,
+        encrypted_bytes,
+    )
+    .map_err(|e| PyTypeError::new_err(format!("Error encrypting block: {e}")))
+}
 /// Symmetrically Decrypt encrypted data in a block.
 #[pyfunction]
 pub fn decrypt_hybrid_block(
     symmetric_key_bytes: Vec<u8>,
-    uid_bytes: Option<Vec<u8>>,
-    block_number: Option<usize>,
+    uid_bytes: Vec<u8>,
+    block_number: usize,
     encrypted_bytes: Vec<u8>,
 ) -> PyResult<Vec<u8>> {
-    //
-    // Check `user_decryption_key_bytes` input param and store it locally
-    if symmetric_key_bytes.len() != 32 {
-        return Err(PyTypeError::new_err("Symmetric key must be 32-bytes long"));
-    }
-
-    //
-    // Check `encrypted_bytes` input param and store it locally
-    if encrypted_bytes.is_empty() {
-        return Err(PyTypeError::new_err("Encrypted value is empty"));
-    }
-
-    //
-    // Parse symmetric key
-    let symmetric_key =
-        <Aes256GcmCrypto as SymmetricCrypto>::Key::try_from_bytes(symmetric_key_bytes)
-            .map_err(|e| PyTypeError::new_err(format!("Deserialize symmetric key failed: {e}")))?;
-
-    //
-    // Parse other input params
-    let uid = uid_bytes.map_or(vec![], |v| v.to_vec());
-    let block_number_value = block_number.unwrap_or(0);
-
-    //
-    // Decrypt block
-    let cleartext =
-        core_decrypt_hybrid_block::<Gpsw<Bls12_381>, Aes256GcmCrypto, MAX_CLEAR_TEXT_SIZE>(
-            &symmetric_key,
-            &uid,
-            block_number_value as usize,
-            &encrypted_bytes,
-        )
-        .map_err(|e| PyTypeError::new_err(format!("Error encrypting block: {e}")))?;
-
-    Ok(cleartext)
+    internal_decrypt_hybrid_block(
+        &symmetric_key_bytes,
+        &uid_bytes,
+        block_number,
+        &encrypted_bytes,
+    )
 }
 
 #[pyfunction]
@@ -228,7 +205,7 @@ pub fn encrypt(
         public_key_bytes,
     )?;
 
-    let ciphertext = encrypt_hybrid_block(header.0, Some(metadata.uid), None, plaintext)?;
+    let ciphertext = encrypt_hybrid_block(header.0, metadata.uid, 0, plaintext)?;
 
     // Encrypted value is composed of: HEADER_LEN (4 bytes) | HEADER | AES_DATA
     let mut encrypted = Vec::<u8>::with_capacity(4 + header.1.len() + ciphertext.len());
@@ -241,14 +218,13 @@ pub fn encrypt(
 #[pyfunction]
 pub fn decrypt(user_decryption_key_bytes: Vec<u8>, encrypted_bytes: Vec<u8>) -> PyResult<Vec<u8>> {
     let header_size = get_encrypted_header_size(encrypted_bytes.clone())?;
-    let header = encrypted_bytes[4..4 + header_size as usize].to_vec();
-    let ciphertext = encrypted_bytes[4 + header_size as usize..].to_vec();
+    let header = &encrypted_bytes[4..4 + header_size as usize];
+    let ciphertext = &encrypted_bytes[4 + header_size as usize..];
 
-    let cleartext_header = decrypt_hybrid_header(user_decryption_key_bytes, header)?;
+    let cleartext_header = internal_decrypt_hybrid_header(&user_decryption_key_bytes, header)?;
 
     let metadata = Metadata::from_bytes(&cleartext_header.1)
         .map_err(|e| PyTypeError::new_err(format!("Error deserializing metadata: {e}")))?;
 
-    let cleartext = decrypt_hybrid_block(cleartext_header.0, Some(metadata.uid), None, ciphertext)?;
-    Ok(cleartext)
+    internal_decrypt_hybrid_block(&cleartext_header.0, &metadata.uid, 0, ciphertext)
 }