+++ src/interfaces/pyo3/mod.rs
@@ -1,7 +1,9 @@
 use pyo3::{pymodule, types::PyModule, wrap_pyfunction, PyResult, Python};
 
 use self::{
-    generate_gpsw_keys::{generate_master_keys, generate_user_private_key, rotate_attributes},
+    generate_gpsw_keys::{
+        generate_master_keys, generate_policy, generate_user_private_key, rotate_attributes,
+    },
     hybrid_gpsw_aes::{
         decrypt, decrypt_hybrid_block, decrypt_hybrid_header, encrypt, encrypt_hybrid_block,
         encrypt_hybrid_header, get_encrypted_header_size,
@@ -20,6 +22,7 @@ impl From<FormatErr> for pyo3::PyErr {
 fn abe_gpsw(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(generate_master_keys, m)?)?;
     m.add_function(wrap_pyfunction!(generate_user_private_key, m)?)?;
+    m.add_function(wrap_pyfunction!(generate_policy, m)?)?;
     m.add_function(wrap_pyfunction!(rotate_attributes, m)?)?;
     m.add_function(wrap_pyfunction!(get_encrypted_header_size, m)?)?;
     m.add_function(wrap_pyfunction!(encrypt_hybrid_header, m)?)?;