+++ src/interfaces/pyo3/tests/test.sh
@@ -12,5 +12,5 @@ init(){
 
 rm -f target/wheels/*.whl
 maturin build --cargo-extra-args="--release --features python"
-pip install target/wheels/*.whl
+pip install --force-reinstall target/wheels/*.whl
 python3 src/interfaces/pyo3/tests/demo.py