+++ src/interfaces/wasm_bindgen/hybrid_gpsw_aes_encryption.rs
@@ -28,7 +28,7 @@ use crate::{
     },
 };
 
-pub const MAX_CLEAR_TEXT_SIZE: usize = 1_usize << 30;
+pub const MAX_CLEAR_TEXT_SIZE: usize = 1 << 30;
 
 type PublicKey = <Gpsw<Bls12_381> as AbeScheme>::MasterPublicKey;